INT-175: Initial intercom setup
