+++ src/app.js
@@ -9,6 +9,8 @@ var QueryOptions = require('./models/query-options');
 var Connection = require('./models/connection');
 var Layout = require('./layout');
 var Statusbar = require('./statusbar');
+var livereload = require('./livereload');
+var intercom = require('./intercom');
 var debug = require('debug')('scout-ui:app');
 
 /**
@@ -77,16 +79,15 @@ var Application = State.extend({
    * We have what we need, we can now start our router and show the appropriate page!
    */
   _onDOMReady: function() {
-    var head = document.getElementsByTagName('head')[0];
-    var script = document.createElement('script');
-    script.type = 'text/javascript';
-    script.src = 'http://localhost:35729/livereload.js';
-    head.appendChild(script);
+    livereload.inject();
+    intercom.inject();
 
     new Layout({
       el: document.querySelector('#application'),
       app: this
     }).render();
+
+    this.router.on('page', intercom.update);
 
     this.router.history.start({
       pushState: false,
@@ -112,6 +113,7 @@ var Application = State.extend({
     this.router.history.navigate(hash, {
       trigger: !options.silent
     });
+    intercom.update();
   }
 });
