+++ src/index.jade
@@ -2,7 +2,7 @@ doctype html
 html(lang='en')
   head
     title MongoDB Enterprise Scout
-    meta(http-equiv="Content-Security-Policy", content="default-src *; script-src 'self' http://localhost:35729; style-src 'self' 'unsafe-inline';")
+    meta(http-equiv="Content-Security-Policy", content="default-src *; script-src 'self' http://localhost:35729 https://widget.intercom.io https://js.intercomcdn.com/; style-src 'self' 'unsafe-inline';")
 
     meta(name='viewport', content='initial-scale=1')
     link(rel='stylesheet', href='index.css')
